changes to int

+++ src/strategy/raids/gruulslair/RaidGruulsLairActions.cpp
@@ -729,8 +729,8 @@ bool GruulTheDragonkillerSpreadRangedAction::Execute(Event event)
     if (!initialPositions.count(bot->GetGUID()))
     {
         auto it = std::find(members.begin(), members.end(), bot);
-        uint32 botIndex = (it != members.end()) ? std::distance(members.begin(), it) : 0;
-        uint32 count = members.size();
+        uint8 botIndex = (it != members.end()) ? std::distance(members.begin(), it) : 0;
+        uint8 count = members.size();
 
         float angle = 2 * M_PI * botIndex / count;
         float radius = minRadius + static_cast<float>(rand()) /
